Simple gui to Run mpv
I'm not a programmer, you can do whatever you want with this code.
This code shall open Mpv, after you copy an URL for the first time and after you use
Ctrl+Shift+X

I had set the directory on my paste
                self.mpv_process = subprocess.Popen(["C:\\Ai\\Video Upscalers\\mpv.exe", url])
so you had to change where your program it is.

Feel free to use or change de code, I just did for my purpose.

And before anything. I'm, not a programar. This literally was my first code.... with a lot of help with Gpt.